refactor: Split CLI into modular commands and add issue state locking
Part 1: Split the "God CLI" into modular commands/ directory
- Reduce cli.py from 1,255 lines to 73 lines (lightweight router)
- Create commands/ directory with specialized modules:
  - _shared.py: Common utilities (console, get_config, styles)
  - hook.py: Fast-startup hook command
  - config.py: Config subcommands (path, init, show, edit)
  - doctor.py: Environment diagnostics
  - initrepo.py: Repository initialization
  - utils.py: Validate, reset, notify commands
  - status.py: Status, sprints, models display
  - init.py: Sprint initialization and planning
  - run.py: Main workflow execution
- Use lazy imports in command modules for faster startup

Part 2: Fix race conditions in issue state management
- Add IssueLock class using existing LockFile utility
- Update IssueService with locking on all read/write operations
- Add atomic_update() context manager for read-modify-write
- Update issue_tools.py with atomic writes (temp file + rename)
- All tool functions now acquire lock before accessing state
