+++ claudesprint/tools/issue_tools.py
@@ -1,21 +1,30 @@
 """Issue tools for CLI - operates on current_issue.json.
 
 Replaces the old handoff_tools module.
+
+This module provides atomic writes and file locking to prevent race conditions
+when multiple processes access current_issue.json concurrently.
 """
 
 import json
+import os
+import tempfile
 from dataclasses import dataclass
 from pathlib import Path
 from typing import Any
 
+from claudesprint.utils.issue_lock import IssueLock, get_issue_lock
+
 # Module-level configuration
 _project_dir: Path | None = None
+_issue_lock: IssueLock | None = None
 
 
 def configure(project_dir: Path) -> None:
-    """Configure the issue tools with project directory."""
-    global _project_dir
+    """Configure the issue tools with project directory and lock."""
+    global _project_dir, _issue_lock
     _project_dir = project_dir
+    _issue_lock = get_issue_lock(project_dir)
 
 
 def _get_issue_path() -> Path:
@@ -25,19 +34,65 @@ def _get_issue_path() -> Path:
     return _project_dir / "current_issue.json"
 
 
+def _get_lock() -> IssueLock:
+    """Get the issue lock instance."""
+    if _issue_lock is None:
+        raise RuntimeError("issue_tools not configured. Call configure() first.")
+    return _issue_lock
+
+
 def _load_issue() -> dict[str, Any]:
-    """Load current_issue.json."""
+    """Load current_issue.json (caller must hold lock)."""
     path = _get_issue_path()
     if not path.exists():
         return {}
     return json.loads(path.read_text())
 
 
-def _save_issue(data: dict[str, Any]) -> None:
-    """Save current_issue.json."""
+def _save_issue_atomic(data: dict[str, Any]) -> None:
+    """Save current_issue.json atomically using temp file + rename.
+
+    Uses a temporary file in the same directory then renames to ensure
+    the write is atomic and won't leave corrupted JSON on crash.
+
+    Caller must hold the lock.
+    """
     path = _get_issue_path()
     path.parent.mkdir(parents=True, exist_ok=True)
-    path.write_text(json.dumps(data, indent=2) + "\n")
+
+    # Write to temp file in same directory (ensures same filesystem for rename)
+    fd, temp_path = tempfile.mkstemp(
+        suffix=".tmp.json",
+        dir=path.parent,
+        prefix=".current_issue_"
+    )
+    temp_file = Path(temp_path)
+    try:
+        # Write content
+        content = json.dumps(data, indent=2) + "\n"
+        os.write(fd, content.encode())
+        os.close(fd)
+        fd = -1  # Mark as closed
+
+        # Atomic rename
+        temp_file.rename(path)
+    finally:
+        # Clean up on failure
+        if fd >= 0:
+            try:
+                os.close(fd)
+            except OSError:
+                pass
+        if temp_file.exists():
+            try:
+                temp_file.unlink()
+            except OSError:
+                pass
+
+
+def _save_issue(data: dict[str, Any]) -> None:
+    """Save current_issue.json atomically (caller must hold lock)."""
+    _save_issue_atomic(data)
 
 
 def _get_log_path() -> Path:
@@ -128,19 +183,20 @@ class ToolResult:
 
 
 def get_issue() -> ToolResult:
-    """Get current issue state."""
+    """Get current issue state with locking."""
     try:
-        data = _load_issue()
-        if not data:
+        with _get_lock().locked():
+            data = _load_issue()
+            if not data:
+                return ToolResult(
+                    success=False,
+                    message="No current_issue.json found",
+                )
             return ToolResult(
-                success=False,
-                message="No current_issue.json found",
+                success=True,
+                message="Current issue loaded",
+                data=data,
             )
-        return ToolResult(
-            success=True,
-            message="Current issue loaded",
-            data=data,
-        )
     except Exception as e:
         return ToolResult(success=False, message=f"Failed to load issue: {e}")
 
@@ -149,23 +205,24 @@ def update_issue(
     goal: str | None = None,
     next_action: str | None = None,
 ) -> ToolResult:
-    """Update current issue fields."""
+    """Update current issue fields with locking."""
     try:
-        data = _load_issue()
-        if not data:
-            return ToolResult(success=False, message="No current_issue.json found")
+        with _get_lock().locked():
+            data = _load_issue()
+            if not data:
+                return ToolResult(success=False, message="No current_issue.json found")
 
-        if goal is not None:
-            data["goal"] = goal
-        if next_action is not None:
-            data["next_action"] = next_action
+            if goal is not None:
+                data["goal"] = goal
+            if next_action is not None:
+                data["next_action"] = next_action
 
-        _save_issue(data)
-        return ToolResult(
-            success=True,
-            message="Issue updated",
-            data={"updated_fields": [k for k, v in [("goal", goal), ("next_action", next_action)] if v is not None]},
-        )
+            _save_issue(data)
+            return ToolResult(
+                success=True,
+                message="Issue updated",
+                data={"updated_fields": [k for k, v in [("goal", goal), ("next_action", next_action)] if v is not None]},
+            )
     except Exception as e:
         return ToolResult(success=False, message=f"Failed to update issue: {e}")
 
@@ -176,90 +233,94 @@ def set_next_step(
     next_action: str | None = None,
     clear_failures: bool = False,
 ) -> ToolResult:
-    """Set the next workflow step."""
+    """Set the next workflow step with locking."""
     try:
-        data = _load_issue()
-        if not data:
-            return ToolResult(success=False, message="No current_issue.json found")
+        with _get_lock().locked():
+            data = _load_issue()
+            if not data:
+                return ToolResult(success=False, message="No current_issue.json found")
 
-        old_step = data.get("step", "")
-        data["step"] = step
+            old_step = data.get("step", "")
+            data["step"] = step
 
-        if goal is not None:
-            data["goal"] = goal
-        if next_action is not None:
-            data["next_action"] = next_action
-        if clear_failures:
-            data["current_failures"] = ""
-            data["retry_count"] = 0
+            if goal is not None:
+                data["goal"] = goal
+            if next_action is not None:
+                data["next_action"] = next_action
+            if clear_failures:
+                data["current_failures"] = ""
+                data["retry_count"] = 0
 
-        _save_issue(data)
-        return ToolResult(
-            success=True,
-            message=f"Step changed: {old_step} -> {step}",
-            data={"old_step": old_step, "new_step": step},
-        )
+            _save_issue(data)
+            return ToolResult(
+                success=True,
+                message=f"Step changed: {old_step} -> {step}",
+                data={"old_step": old_step, "new_step": step},
+            )
     except Exception as e:
         return ToolResult(success=False, message=f"Failed to set step: {e}")
 
 
 def record_change(path: str, summary: str) -> ToolResult:
-    """Record a file change."""
+    """Record a file change with locking."""
     try:
-        data = _load_issue()
-        if not data:
-            return ToolResult(success=False, message="No current_issue.json found")
+        with _get_lock().locked():
+            data = _load_issue()
+            if not data:
+                return ToolResult(success=False, message="No current_issue.json found")
 
-        changes = data.get("changes", [])
-        changes.append({"path": path, "summary": summary})
-        data["changes"] = changes
+            changes = data.get("changes", [])
+            changes.append({"path": path, "summary": summary})
+            data["changes"] = changes
 
-        _save_issue(data)
-        return ToolResult(
-            success=True,
-            message=f"Recorded change: {path}",
-            data={"path": path, "summary": summary},
-        )
+            _save_issue(data)
+            return ToolResult(
+                success=True,
+                message=f"Recorded change: {path}",
+                data={"path": path, "summary": summary},
+            )
     except Exception as e:
         return ToolResult(success=False, message=f"Failed to record change: {e}")
 
 
 def record_failure(failure_message: str, increment_retry: bool = True) -> ToolResult:
-    """Record a failure."""
+    """Record a failure with locking."""
     try:
-        data = _load_issue()
-        if not data:
-            return ToolResult(success=False, message="No current_issue.json found")
+        with _get_lock().locked():
+            data = _load_issue()
+            if not data:
+                return ToolResult(success=False, message="No current_issue.json found")
 
-        data["current_failures"] = failure_message
-        if increment_retry:
-            data["retry_count"] = data.get("retry_count", 0) + 1
+            data["current_failures"] = failure_message
+            if increment_retry:
+                data["retry_count"] = data.get("retry_count", 0) + 1
 
-        _save_issue(data)
-        return ToolResult(
-            success=True,
-            message="Failure recorded",
-            data={"retry_count": data.get("retry_count", 0)},
-        )
+            _save_issue(data)
+            return ToolResult(
+                success=True,
+                message="Failure recorded",
+                data={"retry_count": data.get("retry_count", 0)},
+            )
     except Exception as e:
         return ToolResult(success=False, message=f"Failed to record failure: {e}")
 
 
 def clear_failures() -> ToolResult:
-    """Clear failures and reset retry count."""
+    """Clear failures and reset retry count with locking."""
     try:
-        data = _load_issue()
-        if not data:
-            return ToolResult(success=False, message="No current_issue.json found")
+        with _get_lock().locked():
+            data = _load_issue()
+            if not data:
+                return ToolResult(success=False, message="No current_issue.json found")
 
-        data["current_failures"] = ""
-        data["retry_count"] = 0
+            data["current_failures"] = ""
+            data["retry_count"] = 0
 
-        _save_issue(data)
-        return ToolResult(
-            success=True,
-            message="Failures cleared",
-        )
+            _save_issue(data)
+            return ToolResult(
+                success=True,
+                message="Failures cleared",
+            )
     except Exception as e:
         return ToolResult(success=False, message=f"Failed to clear failures: {e}")
 
@@ -270,7 +331,7 @@ def init_issue(
     goal: str | None = None,
     sprint_path: str | None = None,
 ) -> ToolResult:
-    """Initialize current_issue.json for a selected issue.
+    """Initialize current_issue.json for a selected issue with locking.
 
     This creates or overwrites current_issue.json with fresh state for the selected issue.
     Used by select-issue step to persist the issue selection.
@@ -287,77 +348,78 @@ def init_issue(
     from datetime import datetime, UTC
 
     try:
-        # Try to load existing issue to get sprint_path if not provided
-        if sprint_path is None:
-            existing = _load_issue()
-            if existing and "sprint_path" in existing:
-                sprint_path = existing["sprint_path"]
-            else:
-                # Try to find sprint.json in standard locations
-                if _project_dir is None:
-                    return ToolResult(
-                        success=False,
-                        message="Cannot init: no sprint_path provided and no existing issue to get it from",
-                    )
-                # Walk up to find .claudesprint/sprints/
-                project_root = _project_dir.parent.parent  # .claudesprint/project -> .claudesprint -> project_root
-                sprints_dir = project_root / ".claudesprint" / "sprints"
-                if sprints_dir.exists():
-                    # Find first sprint.json
-                    for sprint_dir in sprints_dir.iterdir():
-                        if sprint_dir.is_dir():
-                            sprint_json = sprint_dir / "sprint.json"
-                            if sprint_json.exists():
-                                sprint_path = str(sprint_json)
-                                break
+        with _get_lock().locked():
+            # Try to load existing issue to get sprint_path if not provided
+            if sprint_path is None:
+                existing = _load_issue()
+                if existing and "sprint_path" in existing:
+                    sprint_path = existing["sprint_path"]
+                else:
+                    # Try to find sprint.json in standard locations
+                    if _project_dir is None:
+                        return ToolResult(
+                            success=False,
+                            message="Cannot init: no sprint_path provided and no existing issue to get it from",
+                        )
+                    # Walk up to find .claudesprint/sprints/
+                    project_root = _project_dir.parent.parent  # .claudesprint/project -> .claudesprint -> project_root
+                    sprints_dir = project_root / ".claudesprint" / "sprints"
+                    if sprints_dir.exists():
+                        # Find first sprint.json
+                        for sprint_dir in sprints_dir.iterdir():
+                            if sprint_dir.is_dir():
+                                sprint_json = sprint_dir / "sprint.json"
+                                if sprint_json.exists():
+                                    sprint_path = str(sprint_json)
+                                    break
 
-        if sprint_path is None:
+            if sprint_path is None:
+                return ToolResult(
+                    success=False,
+                    message="Cannot init: could not determine sprint_path",
+                )
+
+            # Validate issue exists in sprint and is selectable
+            is_valid, error_msg, issue_title = _validate_issue_in_sprint(sprint_path, issue_id)
+            if not is_valid:
+                return ToolResult(
+                    success=False,
+                    message=f"Cannot init issue: {error_msg}",
+                )
+
+            # Create fresh current_issue data
+            timestamp = datetime.now(tz=UTC).strftime("%Y-%m-%dT%H:%M:%SZ")
+            session_id = f"{timestamp}/{step}"
+
+            data = {
+                "schema_version": "2.0",
+                "sprint_path": sprint_path,
+                "issue_id": issue_id,
+                "issue_title": issue_title or "",  # From sprint validation
+                "step": step,
+                "chunk_type": "IMPLEMENT",
+                "goal": goal or f"Begin work on issue {issue_id}",
+                "next_action": f"Read documentation and understand requirements",
+                "context": {},
+                "changes": [],
+                "current_failures": "",
+                "retry_count": 0,
+                "total_iterations": 0,
+                "session_id": session_id,
+                "repo_state": {"head_sha": "", "is_dirty": False},
+                "log_tail": "",
+            }
+
+            # Clear the log file when initializing a new issue to avoid mixing
+            # log entries from different issues
+            _clear_log()
+
+            _save_issue(data)
+
             return ToolResult(
-                success=False,
-                message="Cannot init: could not determine sprint_path",
+                success=True,
+                message=f"Initialized current_issue.json for {issue_id} at step {step}",
+                data={"issue_id": issue_id, "step": step, "sprint_path": sprint_path},
             )
-
-        # Validate issue exists in sprint and is selectable
-        is_valid, error_msg, issue_title = _validate_issue_in_sprint(sprint_path, issue_id)
-        if not is_valid:
-            return ToolResult(
-                success=False,
-                message=f"Cannot init issue: {error_msg}",
-            )
-
-        # Create fresh current_issue data
-        timestamp = datetime.now(tz=UTC).strftime("%Y-%m-%dT%H:%M:%SZ")
-        session_id = f"{timestamp}/{step}"
-
-        data = {
-            "schema_version": "2.0",
-            "sprint_path": sprint_path,
-            "issue_id": issue_id,
-            "issue_title": issue_title or "",  # From sprint validation
-            "step": step,
-            "chunk_type": "IMPLEMENT",
-            "goal": goal or f"Begin work on issue {issue_id}",
-            "next_action": f"Read documentation and understand requirements",
-            "context": {},
-            "changes": [],
-            "current_failures": "",
-            "retry_count": 0,
-            "total_iterations": 0,
-            "session_id": session_id,
-            "repo_state": {"head_sha": "", "is_dirty": False},
-            "log_tail": "",
-        }
-
-        # Clear the log file when initializing a new issue to avoid mixing
-        # log entries from different issues
-        _clear_log()
-
-        _save_issue(data)
-
-        return ToolResult(
-            success=True,
-            message=f"Initialized current_issue.json for {issue_id} at step {step}",
-            data={"issue_id": issue_id, "step": step, "sprint_path": sprint_path},
-        )
     except Exception as e:
         return ToolResult(success=False, message=f"Failed to init issue: {e}")
